Uses `getRecentObject()` to address only one state object
The `getRecentObject()` design pattern allows for only the updated JSON
to be checked during a state change. Before, the state of the entire
object had to be revalidated. Not only is this new design pattern more
efficient, but it also prevents side effect errors. The play/pause bug
is also fixed.